Re-package wheel module in directory.
