+++ crates/pex/src/wheel/mod.rs
@@ -1,0 +1,9 @@
+// Copyright 2026 Pex project contributors.
+// SPDX-License-Identifier: Apache-2.0
+
+mod file;
+mod metadata;
+
+pub(crate) use file::WheelFile;
+pub(crate) use metadata::MetadataReader;
+pub use metadata::WheelMetadata;
